Remove Scribe and do some tidying
