--- public/video-ui/src/components/FormFields/richtext/icons.ts
@@ -1,39 +1,0 @@
-export const icons = {
-  strong: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 112.289 128">
-        <path d="M0 113.404l13.855-3.327V17.913L0 14.597V0h60.566C90.5 0 106.19 10.347 106.19 31.952c0 14.403-6.097 25.125-24 29.375v.923c21.415 2.76 30.1 12.74 30.1 31.212C112.29 114.52 97.52 128 64.27 128H0v-14.596zm56.52-58.74c11.074 0 18.46-4.25 18.46-18.097 0-15.135-6.644-19.192-18.663-19.192H44.883v37.288H56.52zm1.835 55.96c13.855 0 20.863-5.52 20.863-19.94 0-14.597-7.008-20.5-20.672-20.5H44.883v40.44h13.472z"/>
-        </svg>`,
-
-  emphasis: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 91.445 128">
-        <path d="M3.715 113.404l14.758-3.327 22.902-92.164-13.105-3.316L31.965 0h59.48l-3.71 14.597-14.763 3.316-23.098 92.164 13.117 3.327L59.493 128H0l3.715-14.596z"/>
-        </svg>`,
-
-  removeFormatting: `<svg xmlns="http://www.w3.org/2000/svg"viewBox="0 0 128 123.468">
-        <path d="M6.07 84.677l11.582-2.48 17.24-68.957H22.065L13.24 30.754H0L8.825 0h80.68L83.85 30.754H70.885l-.414-17.514H58.474l-17.24 68.955 10.345 2.48-2.76 10.897H3.31l2.76-10.895zM75.463 60.433l52.525 52.525-10.506 10.506L64.957 70.94z"/>
-        <path d="M117.494 60.436l10.503 10.51L75.45 123.46l-10.504-10.51z"/>
-        </svg>`,
-
-  link: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 127.838 128.001">
-        <path d="M78.305 22.448l-11.11-10.912c-15.292-15.34-40.34-15.393-55.68-.09-15.31 15.287-15.366 40.332-.09 55.654l13.41 13.447c15.306 15.34 40.345 15.375 55.685.09l-13.895-13.93c-7.68 7.66-20.2 7.643-27.824-.045L25.376 53.226c-7.664-7.697-7.645-20.18.035-27.85 7.695-7.643 20.18-7.643 27.855.055l6.93 6.94 18.11-9.922z"/><path d="M50.946 107.005l9.465 9.463c15.306 15.35 40.34 15.377 55.673.09 15.32-15.305 15.832-39.84.562-55.172L102.75 47.42c-15.266-15.302-40.32-15.347-55.66-.07L60.98 61.3c7.673-7.662 20.2-7.645 27.825.035l13.395 13.43c7.695 7.713 7.66 20.188-.02 27.83-7.68 7.662-20.172 7.643-27.84-.018l-6.945-6.973-16.45 11.403z"/>
-        </svg>`,
-
-  unlink: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 117.896">
-        <path d="M58.97 20.238l-6.755-6.645c-9.297-9.336-24.527-9.35-33.848-.06-9.324 9.282-9.352 24.524-.055 33.854l8.145 8.168c9.312 9.33 24.527 9.357 33.863.062l-8.45-8.482c-4.67 4.654-12.272 4.633-16.92-.02l-8.165-8.177c-4.66-4.674-4.645-12.275.027-16.924 4.676-4.662 12.277-4.662 16.926.027l4.227 4.218 11.004-6.02zM65.242 94.902l5.758 5.73c9.293 9.343 24.523 9.343 33.848.06 9.31-9.29 9.637-24.225.34-33.55l-8.445-8.487c-9.297-9.324-24.512-9.338-33.852-.04l8.45 8.474c4.676-4.66 12.277-4.647 16.926.026l8.148 8.162c4.676 4.69 4.648 12.277-.027 16.932-4.664 4.654-12.25 4.654-16.914-.02l-4.234-4.23L65.24 94.9zM80.836 32.686l-9.395-4.69L85.44 0l9.392 4.688zm18.47 8.55l-4.2-9.623L123.8 19.117l4.2 9.625zm-56.736 76.66l-9.39-4.69L47.176 85.21l9.394 4.688zM4.195 98.776L0 89.155l28.695-12.496 4.196 9.625z"/>
-        </svg>`,
-
-  undo: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 106.934">
-        <path d="M14.586 98.043v-21.94h5.262v20.942c0 3.744.996 5.203 4.21 5.203 2.866 0 5.208-1.4 7.083-2.513v-23.63h5.267v30.07H32.31l-.593-3.573H31.2c-2.4 2.223-5.267 4.154-9.126 4.154-4.683 0-7.488-2.804-7.488-8.713zm30.182-21.94h4.734v3.802h.588c2.342-2.223 5.613-4.388 9.477-4.388 4.678 0 7.31 2.342 7.31 8.193v22.464h-5.265V85.228c0-3.687-.992-5.203-4.21-5.203-2.866 0-5.505 1.402-7.37 2.576v23.575h-5.264v-30.07zM73.602 92.78V90.14c0-10.82 5.146-14.622 11.523-14.622 3.275 0 4.98 1.173 6.553 2.752h.582s-.582-3.396-.582-5.38v-8.775h5.26v42.06h-3.914l-.525-3.104h-.587c-1.756 2.172-4.15 3.865-7.488 3.865-5.5 0-10.823-3.512-10.823-14.157zm18.19 7.197v-18.08c-1.23-1.346-3.1-2.108-5.665-2.108-4.273 0-7.14 2.398-7.14 10.183v2.398c0 8.77 3.214 10.354 7.07 10.354 2.754 0 4.504-1.408 5.734-2.747zm12.052-7.55V90.14c0-9.47 4.965-14.622 12.105-14.622 7.548 0 12.05 4.975 12.05 14.45v2.224c0 9.59-4.97 14.68-12.11 14.68-7.544 0-12.046-5.027-12.046-14.447zm18.84.233V89.91c0-8.017-2.986-10.353-6.793-10.353-4.21 0-6.665 3.157-6.665 10.178v2.752c0 8.07 2.984 10.354 6.785 10.354 4.15.002 6.674-3.04 6.674-10.182zM36.854 41.375H2.197L0 45.03 16.898 61.93l4.788.197 17.337-17.343z"/><path d="M49.072 0C29.225 0 13.08 16.15 13.08 35.996v11.538h12.463V35.996c0-12.96 10.562-23.523 23.53-23.523 12.976 0 23.542 10.562 23.542 23.523v19.358h12.463V35.996C85.078 16.15 68.924 0 49.072 0z"/>
-        </svg>`,
-
-  redo: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 105.214">
-        <path d="M0 75.112h4.625v4.284h.566c2.345-2.91 5.487-4.624 9.44-4.624v4.97c-3.667-.118-7.49 1.084-9.487 2.625v22.106H0v-29.36zM19.014 90.82v-2.003c0-9.076 4.62-14.278 11.43-14.278 7.812 0 10.328 5.657 10.328 13.364v3.032H23.98c.06 7.534 2.92 10.05 8.116 10.05 3.367 0 5.486-.74 7.768-1.998v3.884c-1.54 1.03-4.23 2.28-8.568 2.28-8.114 0-12.282-5.307-12.282-14.333zm4.965-2.682c1.206-.233 3.376-.63 5.6-.63h6.345c0-6.226-1.664-9.024-5.547-9.024-3.65 0-6.34 2.514-6.4 9.654zm22.11 3.255v-2.575c0-10.567 5.023-14.278 11.25-14.278 3.2 0 4.854 1.146 6.398 2.688h.568s-.568-3.316-.568-5.253v-8.57h5.137v41.07h-3.822l-.518-3.03h-.572c-1.71 2.118-4.05 3.77-7.307 3.77-5.375 0-10.566-3.428-10.566-13.822zm17.762 7.027V80.766c-1.197-1.313-3.033-2.06-5.543-2.06-4.17 0-6.968 2.343-6.968 9.943v2.342c0 8.563 3.145 10.11 6.912 10.11 2.676 0 4.4-1.373 5.598-2.682zm11.76-7.372v-2.23c0-9.25 4.85-14.278 11.825-14.278 7.367 0 11.773 4.857 11.773 14.11v2.17c0 9.365-4.86 14.334-11.828 14.334-7.372 0-11.77-4.91-11.77-14.106zm18.392.228v-2.683c0-7.828-2.91-10.11-6.62-10.11-4.124 0-6.513 3.083-6.513 9.938v2.688c0 7.88 2.907 10.11 6.618 10.11 4.063 0 6.516-2.97 6.516-9.942zM125.86 40.39H92.017l-2.114 3.327 16.92 16.934 4.68-.18L128 43.96z"/><path d="M80.086 0c19.38 0 35.146 15.78 35.146 35.143v11.28h-12.178v-11.28c0-12.66-10.303-22.963-22.97-22.963S57.1 22.482 57.1 35.143V54.05h-12.17V35.143C44.928 15.78 60.708 0 80.086 0z"/>
-        </svg>`,
-
-  heading: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 72.791">
-        <path d="M0 64.62l7.656-1.86V11.175L0 9.31V1.143h32.258V9.31l-6.922 1.756v20.37H49.22v-20.37l-7.134-1.758V1.144H74.34V9.31l-7.543 1.864V62.76l7.652 1.86v8.17H42.085v-8.17l7.133-1.758V40.845H25.335v22.018l6.922 1.757v8.172H0V64.62zm77.445-1.14L94.19 47.044c8.38-8.272 13.86-14.788 13.86-22.853 0-7.337-4.04-11.268-13.03-11.268-5.278 0-10.65 1.144-15.095 2.692v-9.72C85.098 3.002 93.363 0 102.152 0c15.2 0 24.19 7.754 24.19 21.606 0 11.275-8.27 20.477-21.71 30.917l-10.03 7.963v.2h20.68l3.827-8.062H128l-1.547 20.167H77.445v-9.31z"/>
-        </svg>`,
-
-  cleanUp: `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 114.189">
-        <path d="M12.285,114.189L0,100.128l77.024-67.306l12.281,14.062L12.285,114.189z M6.6,100.572l6.131,7.019l69.978-61.151
-    l-6.128-7.019L6.6,100.572z"/><path d="M26.826 98.371L17.621 87.837 62.398 48.711 71.607 59.251zM125.753 41.614L120.434 51.311 128 59.362 117.137 57.312 111.841 66.989 110.439 56.016 99.57 53.945 109.58 49.262 108.182 38.293 115.76 46.349zM122.693 5.688L115.926 18.035 125.547 28.277 111.727 25.668 104.988 37.984 103.195 24.022 89.363 21.388 102.111 15.421 100.328 1.459 109.977 11.721zM99.259 63.85L94.027 70.109 98.355 76.994 90.789 73.955 85.588 80.205 86.146 72.078 78.582 69.044 86.484 67.075 87.018 58.948 91.357 65.844zM55.137 23.941L53.676 31.766 60.653 35.592 52.748 36.602 51.289 44.44 47.879 37.232 39.977 38.247 45.777 32.785 42.349 25.582 49.346 29.419zM81.58 0L78.756 10.51 87.861 16.431 76.988 16.99 74.176 27.49 70.283 17.339 59.432 17.874 67.895 11.065 64.004.908 73.12 6.84z"/>
-        </svg>`
-};
