+++ public/video-ui/src/components/FormFields/richtext/utils/listsHelpers.ts
