+++ public/video-ui/styles/components/_details-list.scss
@@ -18,7 +18,7 @@
   }
 }
 
-.details-list__field--scribe > p,ul {
+.details-list__field-with-content > p,ul {
   margin: 0;
 }
