--- public/video-ui/styles/components/_scribe.scss
@@ -1,42 +1,0 @@
-.scribe__toolbar {
-  width: 100%;
-}
-
-.scribe__toolbar__item {
-  padding: 10px;
-  font-size: 12px;
-  background-color: $color650Grey;
-  color: $color300Grey;
-  border: {
-    left: 1px solid $color500Grey;
-    top: 1px solid $color500Grey;
-  }
-
-  &:last-child {
-    border-right: 1px solid $color500Grey;
-  }
-
-  &:hover {
-    background-color: $color625Grey;
-    color: $brandColor;
-  }
-
-  i.icon {
-    font-size: $iconSize;
-  }
-}
-
-.scribe__row {
-  display: flex;
-}
-
-.scribe__container {
-  width: 100%;
-}
-
-.scribe__editor {
-  width: 100%;
-  padding: 20px;
-  background: $color650Grey;
-  border: 1px solid $color500Grey;
-}
