+++ public/video-ui/styles/main.scss
@@ -45,7 +45,6 @@
   'components/usage',
   'components/expiry-date',
   'components/advanced',
-  'components/scribe',
   'components/prosemirror',
   'components/presence',
   'components/scheduledLaunch',
